+++ packages/api-reference/src/features/Operation/components/callbacks/Callback.vue
@@ -1,44 +1,38 @@
 <script setup lang="ts">
+import type { HttpMethod as HttpMethodType } from '@scalar/helpers/http/http-methods'
 import { ScalarIconCaretRight } from '@scalar/icons'
-import { requestSchema, type Collection } from '@scalar/oas-utils/entities/spec'
-import { schemaModel } from '@scalar/oas-utils/helpers'
-import type { OpenAPIV3_1 } from '@scalar/types/legacy'
-import { computed } from 'vue'
+import { type OperationObject } from '@scalar/workspace-store/schemas/v3.1/strict/path-operations'
+import { type Dereference } from '@scalar/workspace-store/schemas/v3.1/type-guard'
 
 import { HttpMethod } from '@/components/HttpMethod'
 import OperationParameters from '@/features/Operation/components/OperationParameters.vue'
 import OperationResponses from '@/features/Operation/components/OperationResponses.vue'
 import type { Schemas } from '@/features/Operation/types/schemas'
 
-const { callback, collection, method, name, schemas, url } = defineProps<{
-  callback: OpenAPIV3_1.OperationObject
-  collection: Collection
-  method: string
+const { method, name, schemas, url, path, operationMethod } = defineProps<{
+  callback: Dereference<OperationObject>
+  method: HttpMethodType
+  path: string
+  operationMethod: HttpMethodType
   name: string
   schemas?: Schemas
   url: string
 }>()
-
-/** This should get us 90% of the way there, will fix the rest on new store */
-const operation = computed(() =>
-  schemaModel({ ...callback, path: url, method }, requestSchema, false),
-)
 </script>
 
 <template>
-  <details
-    v-if="collection && operation"
-    class="group">
+  <details class="group callback-list-item">
     <!-- Title -->
     <summary
-      class="font-code bg-b-1 callback-sticky-offset sticky flex cursor-pointer flex-row items-center gap-2 border-t py-3 text-sm group-open:flex-wrap">
+      class="font-code bg-b-1 callback-sticky-offset callback-list-item-title sticky flex cursor-pointer flex-row items-center gap-2 border-t py-3 text-sm group-open:flex-wrap">
       <ScalarIconCaretRight
-        class="text-c-3 group-hover:text-c-1 absolute -left-5 size-4 transition-transform group-open:rotate-90" />
+        class="callback-list-item-icon text-c-3 group-hover:text-c-1 absolute -left-5 size-4 transition-transform group-open:rotate-90" />
       <HttpMethod
         as="span"
         class="request-method"
         :method="method" />
-      <div class="text-c-1 truncate leading-3 group-open:whitespace-normal">
+      <div
+        class="text-c-1 min-w-0 flex-1 truncate leading-3 group-open:whitespace-normal">
         {{ name }}
         <span class="text-c-2">
           {{ url }}