+++ packages/api-reference/src/features/Operation/layouts/ClassicLayout.vue
@@ -1,22 +1,26 @@
 <script setup lang="ts">
 import {
-  ScalarIcon,
+  ScalarErrorBoundary,
   ScalarIconButton,
   ScalarMarkdown,
 } from '@scalar/components'
-import { ScalarIconWebhooksLogo } from '@scalar/icons'
-import type {
-  Collection,
-  Request,
-  Server,
-} from '@scalar/oas-utils/entities/spec'
+import type { HttpMethod as HttpMethodType } from '@scalar/helpers/http/http-methods'
+import {
+  ScalarIconCopy,
+  ScalarIconPlay,
+  ScalarIconWebhooksLogo,
+} from '@scalar/icons'
 import {
   getOperationStability,
   getOperationStabilityColor,
   isOperationDeprecated,
 } from '@scalar/oas-utils/helpers'
-import type { OpenAPIV3_1, XScalarStability } from '@scalar/types/legacy'
+import type { OpenAPIV3_1 } from '@scalar/openapi-types'
 import { useClipboard } from '@scalar/use-hooks/useClipboard'
+import type { OperationObject } from '@scalar/workspace-store/schemas/v3.1/strict/path-operations'
+import type { SecuritySchemeObject } from '@scalar/workspace-store/schemas/v3.1/strict/security-scheme'
+import type { ServerObject } from '@scalar/workspace-store/schemas/v3.1/strict/server'
+import type { Dereference } from '@scalar/workspace-store/schemas/v3.1/type-guard'
 import { computed } from 'vue'
 
 import { Anchor } from '@/components/Anchor'
@@ -24,35 +28,32 @@ import { Badge } from '@/components/Badge'
 import { HttpMethod } from '@/components/HttpMethod'
 import OperationPath from '@/components/OperationPath.vue'
 import { SectionAccordion } from '@/components/Section'
-import { ExampleRequest } from '@/features/example-request'
 import { ExampleResponses } from '@/features/example-responses'
+import Callbacks from '@/features/Operation/components/callbacks/Callbacks.vue'
+import OperationParameters from '@/features/Operation/components/OperationParameters.vue'
+import OperationResponses from '@/features/Operation/components/OperationResponses.vue'
 import type { Schemas } from '@/features/Operation/types/schemas'
 import { TestRequestButton } from '@/features/test-request-button'
 import { useConfig } from '@/hooks/useConfig'
+import { RequestExample } from '@/v2/blocks/scalar-request-example-block'
+import { useStore } from '@/v2/hooks/useStore'
 
-import OperationParameters from '../components/OperationParameters.vue'
-import OperationResponses from '../components/OperationResponses.vue'
-
-const { request, operation, path, isWebhook } = defineProps<{
+const { operation, path, isWebhook } = defineProps<{
   id: string
   path: string
-  method: OpenAPIV3_1.HttpMethods
-  operation: OpenAPIV3_1.OperationObject<{
-    'x-scalar-stability': XScalarStability
-  }>
+  method: HttpMethodType
+  operation: Dereference<OperationObject>
+  oldOperation: OpenAPIV3_1.OperationObject
   isWebhook: boolean
-  /**
-   * @deprecated Use `document` instead
-   */
-  collection: Collection
-  server: Server | undefined
-  request: Request | undefined
+  server: ServerObject | undefined
+  securitySchemes: SecuritySchemeObject[]
   schemas?: Schemas
 }>()
 
-const operationTitle = computed(() => operation?.summary || path || '')
+const operationTitle = computed(() => operation.summary || path || '')
 
 const { copyToClipboard } = useClipboard()
+const { workspace } = useStore()
 const config = useConfig()
 
 const emit = defineEmits<{
@@ -105,26 +106,25 @@ const handleDiscriminatorChange = (type: string) => {
     </template>
     <template #actions="{ active }">
       <TestRequestButton
-        v-if="active && request"
-        :operation="request" />
-      <ScalarIcon
+        v-if="active && !isWebhook"
+        :method="method"
+        :path="path" />
+      <ScalarIconPlay
         v-else-if="!config?.hideTestRequestButton"
-        class="endpoint-try-hint size-6"
-        icon="Play"
-        thickness="1.75px" />
+        class="endpoint-try-hint size-4.5" />
       <ScalarIconButton
         class="endpoint-copy p-0.5"
-        icon="Clipboard"
+        :icon="ScalarIconCopy"
         label="Copy endpoint URL"
         size="xs"
         variant="ghost"
         @click.stop="copyToClipboard(path)" />
     </template>
     <template
-      v-if="operation?.description"
+      v-if="operation.description"
       #description>
       <ScalarMarkdown
-        :value="operation?.description"
+        :value="operation.description"
         withImages
         withAnchors
         transformType="heading"
@@ -134,26 +134,48 @@ const handleDiscriminatorChange = (type: string) => {
       <div class="operation-details-card">
         <div class="operation-details-card-item">
           <OperationParameters
-            :parameters="operation?.parameters"
-            :requestBody="operation?.requestBody"
+            :parameters="oldOperation.parameters"
+            :requestBody="oldOperation.requestBody"
             :schemas="schemas"
             @update:modelValue="handleDiscriminatorChange" />
         </div>
         <div class="operation-details-card-item">
           <OperationResponses
             :collapsableItems="false"
-            :responses="operation.responses"
+            :responses="oldOperation.responses"
+            :schemas="schemas" />
+        </div>
+
+        <!-- Callbacks -->
+        <div
+          v-if="operation?.callbacks"
+          class="operation-details-card-item">
+          <Callbacks
+            :method="method"
+            :path="path"
+            :callbacks="operation.callbacks"
             :schemas="schemas" />
         </div>
       </div>
-      <ExampleResponses :responses="operation.responses" />
-      <ExampleRequest
-        :request="request"
-        :method="method"
-        :collection="collection"
-        :operation="operation"
-        :server="server"
-        @update:modelValue="handleDiscriminatorChange" />
+
+      <ExampleResponses
+        class="operation-example-card"
+        :responses="operation.responses" />
+
+      <!-- New Example Request -->
+      <ScalarErrorBoundary>
+        <RequestExample
+          class="operation-example-card"
+          :method="method"
+          :selectedServer="server"
+          :selectedClient="workspace['x-scalar-default-client']"
+          :securitySchemes="securitySchemes"
+          :path="path"
+          fallback
+          :operation="operation"
+          :schemas="schemas"
+          @update:modelValue="handleDiscriminatorChange" />
+      </ScalarErrorBoundary>
     </div>
   </SectionAccordion>
 </template>
@@ -202,7 +224,7 @@ const handleDiscriminatorChange = (type: string) => {
   background: currentColor;
   opacity: 0.15;
 
-  border-radius: var(--scalar-radius-lg);
+  border-radius: var(--scalar-radius);
 }
 
 .endpoint-anchor {
@@ -282,16 +304,18 @@ const handleDiscriminatorChange = (type: string) => {
 }
 
 .endpoint-content > * {
-  max-height: unset;
+  min-width: 0;
 }
 
 .operation-details-card {
   display: flex;
   flex-direction: column;
   gap: 12px;
+  min-width: 0;
 }
-.operation-details-card-item :deep(.parameter-list) {
-  border: 1px solid var(--scalar-border-color);
+.operation-details-card-item :deep(.parameter-list),
+.operation-details-card-item :deep(.callbacks-list) {
+  border: var(--scalar-border-width) solid var(--scalar-border-color);
   border-radius: var(--scalar-radius-lg);
   margin-top: 0;
 }
@@ -313,14 +337,15 @@ const handleDiscriminatorChange = (type: string) => {
 }
 .operation-details-card :deep(.parameter-item) {
   margin: 0;
-  padding: 0 9px;
+  padding: 0;
 }
 .operation-details-card :deep(.property) {
   padding: 9px;
   margin: 0;
 }
 .operation-details-card :deep(.parameter-list-title),
-.operation-details-card :deep(.request-body-title) {
+.operation-details-card :deep(.request-body-title),
+.operation-details-card :deep(.callbacks-title) {
   text-transform: uppercase;
   font-weight: var(--scalar-bold);
   font-size: var(--scalar-mini);
@@ -330,16 +355,42 @@ const handleDiscriminatorChange = (type: string) => {
   margin: 0;
 }
 
+.operation-details-card :deep(.callback-list-item-title) {
+  padding-left: 28px;
+  padding-right: 12px;
+}
+
+.operation-details-card :deep(.callback-list-item-icon) {
+  left: 6px;
+}
+
+.operation-details-card :deep(.callback-operation-container) {
+  padding-inline: 9px;
+  padding-bottom: 9px;
+}
+
+.operation-details-card :deep(.callback-operation-container > .request-body),
+.operation-details-card :deep(.callback-operation-container > .parameter-list) {
+  border: none;
+}
+
+.operation-details-card
+  :deep(.callback-operation-container > .request-body > .request-body-header) {
+  padding: 0;
+  padding-bottom: 9px;
+  border-bottom: var(--scalar-border-width) solid var(--scalar-border-color);
+}
+
 .operation-details-card :deep(.request-body-description) {
   margin-top: 0;
   padding: 9px 9px 0 9px;
-  border-top: 1px solid var(--scalar-border-color);
+  border-top: var(--scalar-border-width) solid var(--scalar-border-color);
 }
 
 .operation-details-card :deep(.request-body) {
   margin-top: 0;
   border-radius: var(--scalar-radius-lg);
-  border: 1px solid var(--scalar-border-color);
+  border: var(--scalar-border-width) solid var(--scalar-border-color);
 }
 
 .operation-details-card :deep(.request-body-header) {
@@ -351,17 +402,29 @@ const handleDiscriminatorChange = (type: string) => {
   margin-right: 9px;
 }
 
-.operation-details-card :deep(.request-body-schema > .schema-card) {
-  border-radius: var(--scalar-radius-lg);
-  border: 1px solid var(--scalar-border-color);
-  margin: 9px;
+.operation-details-card
+  :deep(.schema-card--open + .schema-card:not(.schema-card--open)) {
+  margin-inline: 9px;
+  margin-bottom: 9px;
 }
-
 .operation-details-card :deep(.request-body-schema .property--level-0) {
   padding: 0;
 }
 
 .operation-details-card :deep(.selected-content-type) {
   margin-right: 9px;
+}
+
+.operation-example-card {
+  position: sticky;
+  top: calc(var(--refs-header-height) + 24px);
+  max-height: calc(((var(--full-height) - var(--refs-header-height)) - 48px));
+}
+
+@media (max-width: 600px) {
+  .operation-example-card {
+    max-height: unset;
+    position: static;
+  }
 }
 </style>