+++ integrations/fastify/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @scalar/fastify-api-reference
 
+## 1.32.9
+
+## 1.32.8
+
+### Patch Changes
+
+- Updated dependencies [591562f]
+  - @scalar/openapi-types@0.3.6
+  - @scalar/openapi-parser@0.18.1
+  - @scalar/core@0.3.9
+
+## 1.32.7
+
 ## 1.32.6
 
 ### Patch Changes