+++ integrations/hono/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @scalar/hono-api-reference
 
+## 0.9.11
+
+### Patch Changes
+
+- @scalar/core@0.3.9
+
 ## 0.9.10
 
 ### Patch Changes