+++ packages/mock-server/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @scalar/mock-server
 
+## 0.5.17
+
+### Patch Changes
+
+- Updated dependencies [591562f]
+  - @scalar/openapi-types@0.3.6
+  - @scalar/oas-utils@0.4.14
+  - @scalar/openapi-parser@0.18.1
+
+## 0.5.16
+
+### Patch Changes
+
+- 3e657c9: feat(mock-server): update galaxy.scalar.com to use default theme
+- Updated dependencies [71958fd]
+  - @scalar/oas-utils@0.4.13
+
 ## 0.5.15
 
 ### Patch Changes