+++ packages/oas-utils/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @scalar/oas-utils
 
+## 0.4.14
+
+### Patch Changes
+
+- Updated dependencies [591562f]
+- Updated dependencies [3f2ea8a]
+- Updated dependencies [828c894]
+- Updated dependencies [b5bcce7]
+- Updated dependencies [8a67f4f]
+  - @scalar/openapi-types@0.3.6
+  - @scalar/types@0.2.9
+  - @scalar/workspace-store@0.10.1
+  - @scalar/themes@0.13.10
+
+## 0.4.13
+
+### Patch Changes
+
+- 71958fd: fix: include port in server URL
+- Updated dependencies [a85480e]
+- Updated dependencies [8b9c48e]
+  - @scalar/workspace-store@0.10.0
+
 ## 0.4.12
 
 ### Patch Changes