+++ packages/postman-to-openapi/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @scalar/postman-to-openapi
 
+## 0.3.17
+
+### Patch Changes
+
+- Updated dependencies [591562f]
+  - @scalar/openapi-types@0.3.6
+  - @scalar/oas-utils@0.4.14
+
+## 0.3.16
+
+### Patch Changes
+
+- Updated dependencies [71958fd]
+  - @scalar/oas-utils@0.4.13
+
 ## 0.3.15
 
 ### Patch Changes